refactor(utils): extract hashing helpers to dedicated module
- Move sha256, sha256_cbor, and get_hash_fn_by_name from vllm.utils.__init__
  to vllm/utils/hashing.py
- Update import in vllm/v1/engine/core.py to use new module path
- Add vllm/utils/hashing.py to pickle import allowlist
- Use direct 'import pickle' instead of importlib workaround

These hashing functions are internal-only, so no backward compatibility
is needed. This refactoring improves code organization while keeping
the implementation cleaner.
