+++ vllm/v1/core/kv_cache_utils.py
@@ -12,7 +12,8 @@ from typing import Any, NewType, TypeAlias
 from vllm import envs
 from vllm.config import VllmConfig
 from vllm.logger import init_logger
-from vllm.utils import cdiv, sha256_cbor
+from vllm.utils import cdiv
+from vllm.utils.hashing import sha256_cbor
 from vllm.utils.mem_constants import GiB_bytes
 from vllm.v1.kv_cache_interface import (
     ChunkedLocalAttentionSpec,
